FEN ep-square modification
modify FEN string to only include en passant square if
an enemy pawn is next to it, does not check legality of capture.
the idea is domain reduction (reduce duplicate positions). this
varies from the FEN standard but is a common implementation

note: first use of Go generics in this commit :D